+++ src/llm/gemini_validation.py
@@ -7,7 +7,7 @@ class GeminiConfig(BaseModel):
     gemini_api_key: str = Field(..., min_length=1, description="Klucz API Google Gemini")
     temperature: Optional[float] = Field(None, description="Wartość parametru temperature")
     top_p: Optional[float] = Field(None, description="Wartość parametru top_p")
-    top_k: Optional[float] = Field(None, description="Wartość parametru top_k")
+    top_k: Optional[int] = Field(None, description="Wartość parametru top_k")
     
     @validator('gemini_api_key')
     def validate_api_key(cls, v):
